azure_auth_core update
