+++ msticpy/data/azure_data.py
@@ -6,18 +6,12 @@
 """Uses the Azure Python SDK to collect and return details related to Azure."""
 from typing import Optional, Dict, Tuple
 import datetime
-import os
 
 import attr
 import pandas as pd
 import numpy as np
 
 
-from msrest.authentication import BasicTokenAuthentication
-from azure.core.pipeline.policies import BearerTokenCredentialPolicy
-from azure.core.pipeline import PipelineRequest, PipelineContext
-from azure.core.pipeline.transport import HttpRequest
-from azure.identity import DefaultAzureCredential
 from azure.mgmt.subscription import SubscriptionClient
 from azure.mgmt.resource import ResourceManagementClient
 from azure.mgmt.network import NetworkManagementClient
@@ -27,13 +21,14 @@ from azure.mgmt.compute.models import VirtualMachineInstanceView
 from azure.common.exceptions import CloudError
 
 from ..common.azure_auth import az_connect
+from ..common.azure_auth_core import AzCredentials
 
 from ..common.exceptions import (
     MsticpyAzureConfigError,
     MsticpyNotConnectedError,
     MsticpyResourceException,
 )
-from ..common.provider_settings import get_provider_settings
+
 from .._version import VERSION
 
 __version__ = VERSION
@@ -97,9 +92,6 @@ class InterfaceItems:
     subnet_nsg = attr.ib()
     subnet_route_table = attr.ib()
 
-# uncomment once circular dependencies fixed
-# pylint: enable=too-few-public-methods, too-many-instance-attributes
-
 
 class AzureData:
     """Class for returning data on an Azure tenant."""
@@ -107,7 +99,7 @@ class AzureData:
     def __init__(self, connect: bool = False):
         """Initialize connector for Azure Python SDK."""
         self.connected = False
-        self.credentials: Optional[DefaultAzureCredential] = None
+        self.credentials: Optional[AzCredentials] = None
         self.sub_client: Optional[SubscriptionClient] = None
         self.resource_client: Optional[ResourceManagementClient] = None
         self.network_client: Optional[NetworkManagementClient] = None
@@ -116,39 +108,21 @@ class AzureData:
         if connect is True:
             self.connect()
 
-    def connect(self, client_id: str = None, tenant_id: str = None, secret: str = None):
+    def connect(
+        self,
+        client_id: str = None,
+        tenant_id: str = None,
+        secret: str = None,
+        silent: bool = False,
+    ):
         """Authenticate with the SDK."""
-        # Use details of msticpyyaml if not provided
-        #if client_id is None and tenant_id is None and secret is None:
-            #data_provs = get_provider_settings(config_section="DataProviders")
-            #az_cli_config = data_provs.get("AzureCLI")
-            # az_cli_config = config.settings.get("AzureCLI")
-            # if not az_cli_config:
-            #  raise MsticpyAzureConfigError(
-            #     "No AzureCLI section found in configuration settings.",
-            #     title="no AzureCLI settings available.",
-            # )
-
-            #try:
-            #    config_items = az_cli_config.args
-            #    os.environ["AZURE_CLIENT_ID"] = config_items["clientId"]
-            #    os.environ["AZURE_TENANT_ID"] = config_items["tenantId"]
-            #    os.environ["AZURE_CLIENT_SECRET"] = config_items["clientSecret"]
-            #except:
-            #    pass
-                # KeyError as key_err:
-                # key_name = key_err.args[0]
-                # raise MsticpyAzureConfigError(
-                #    f"{key_name} is missing from AzureCLI section in your",
-                #    "configuration.",
-                #    title="missing f{key_name} settings for AzureCLI.",
-                # ) from key_err
-        # Create credentials and connect to the subscription client to validate
-        self.credentials = az_connect()
+        self.credentials = az_connect(
+            client_id=client_id, tenant_id=tenant_id, secret=secret, silent=silent
+        )
         if not self.credentials:
             raise CloudError("Could not obtain credentials.")
         self._check_client("sub_client")
-        # self.sub_client = SubscriptionClient(CredentialWrapper(self.credentials))
+        self.sub_client = SubscriptionClient(self.credentials.legacy)
         if not self.sub_client:
             raise CloudError("Could not create a Subscription client.")
         self.connected = True
@@ -657,13 +631,23 @@ class AzureData:
         start = datetime.datetime.now().date()
         end = start - datetime.timedelta(days=start_time)
 
-        mon_details = self.monitoring_client.metrics.list(  # type: ignore
-            resource_id,
-            timespan=f"{end}/{start}",
-            interval=interval,
-            metricnames=f"{metrics}",
-            aggregation="Total",
-        )
+        try:
+            mon_details = self.monitoring_client.metrics.list(  # type: ignore
+                resource_id,
+                timespan=f"{end}/{start}",
+                interval=interval,
+                metricnames=f"{metrics}",
+                aggregation="Total",
+            )
+        except AttributeError:
+            self._legacy_auth("monitoring_client", sub_id)
+            mon_details = self.monitoring_client.metrics.list(  # type: ignore
+                resource_id,
+                timespan=f"{end}/{start}",
+                interval=interval,
+                metricnames=f"{metrics}",
+                aggregation="Total",
+            )
         results = {}
         # Create a dict of all the results returned
         for metric in mon_details.value:
@@ -766,59 +750,3 @@ class AzureData:
             setattr(self, client_name, client(self.credentials.legacy))
         else:
             setattr(self, client_name, client(self.credentials.legacy, sub_id))
-
-
-# Class to extract v1 authentication token from DefaultAzureCredential object.
-# Credit - https://gist.github.com/lmazuel/cc683d82ea1d7b40208de7c9fc8de59d
-class CredentialWrapper(BasicTokenAuthentication):
-    """Class for handling legacy auth conversion."""
-
-    def __init__(
-        self,
-        credential=None,
-        resource_id="https://management.azure.com/.default",
-        **kwargs,
-    ):
-        """
-        Wrap any azure-identity credential to work with SDK that needs azure.common.credentials.
-
-        Parameters
-        ----------
-        credential : [type], optional
-            Any azure-identity credential, by default DefaultAzureCredential
-        resource_id : str, optional
-            The scope to use to get the token, by default "https://management.azure.com/.default"
-
-        """
-        super(CredentialWrapper, self).__init__(None)
-        if credential is None:
-            credential = DefaultAzureCredential()
-        self._policy = BearerTokenCredentialPolicy(credential, resource_id, **kwargs)
-
-    def set_token(self):
-        """
-        Ask the azure-core BearerTokenCredentialPolicy policy to get a token.
-
-        Using the policy gives us for free the caching system of azure-core.
-        We could make this code simpler by using private method, but by definition
-        I can't assure they will be there forever, so mocking a fake call to the policy
-        to extract the token, using 100% public API.
-
-        """
-        request = _make_request()
-        self._policy.on_request(request)
-        # Read Authorization, and get the second part after Bearer
-        token = request.http_request.headers["Authorization"].split(" ", 1)[1]
-        self.token = {"access_token": token}
-
-    def signed_session(self, session=None):
-        """Wrap signed session object."""
-        self.set_token()
-        return super(CredentialWrapper, self).signed_session(session)
-
-
-def _make_request():
-    """Make mocked request to get token."""
-    return PipelineRequest(
-        HttpRequest("CredentialWrapper", "https://fakeurl"), PipelineContext(None)
-    )
